requirements.txt
allure-pytest==2.13.2
allure-python-commons==2.13.2
attrs==23.2.0
behave==1.2.6
blinker==1.6.3
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
cowsay==6.1
data==0.4
decorator==5.1.1
enum34==1.1.10
Flask==3.0.0
funcsigs==1.0.2
h11==0.14.0
idna==3.6
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
lxml==5.1.0
MarkupSafe==2.1.3
namedlist==1.7
outcome==1.3.0.post0
packaging==23.2
parse==1.20.1
parse-type==0.6.2
pluggy==1.4.0
psycopg2-binary==2.9.9
PySocks==1.7.1
pyTelegramBotAPI==4.14.1
pytest==8.0.0
python-dotenv==1.0.0
requests==2.31.0
selenium==4.16.0
setuptools==69.1.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
trio==0.24.0
trio-websocket==0.11.1
urllib3==2.1.0
webdriver-manager==4.0.1
Werkzeug==3.0.1
wheel==0.42.0
wsproto==1.2.0
